[#1573] initial timeseries file type models
