+++ hs_file_types/models/__init__.py
@@ -4,3 +4,5 @@ from raster import GeoRasterFileMetaData, GeoRasterLogicalFile  # noqa
 from netcdf import NetCDFFileMetaData, NetCDFLogicalFile        # noqa
 from geofeature import GeoFeatureFileMetaData, GeoFeatureLogicalFile    # noqa
 from reftimeseries import RefTimeseriesFileMetaData, RefTimeseriesLogicalFile   # noqa
+from timseries import TimeSeriesFileMetaData, TimeSeriesLogicalFile     # noqa
+
